Tidy up Policy base classes
+ Establish PolicyBase for functionality common to Policy and
Policy<TResult>
+ Esablish separate partial class files for the internals of Policy
operation, from the public Execute (etc) overloads
