+++ src/Polly.Shared/IAsyncPolicy.TResult.cs
