+++ src/Polly.Shared/ISyncPolicy.TResult.cs
