+++ src/Polly.Shared/Policy.ExecuteOverloads.cs
@@ -1,0 +1,594 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Threading;
+
+namespace Polly
+{
+    public abstract partial class Policy : ISyncPolicy
+    {
+        #region Execute overloads
+
+        /// <summary>
+        /// Executes the specified action within the policy.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        [DebuggerStepThrough]
+        public void Execute(Action action)
+        {
+            Execute((ctx, ct) => action(), new Context(), DefaultCancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
+        [DebuggerStepThrough]
+        public void Execute(Action action, IDictionary<string, object> contextData)
+        {
+            Execute((ctx, ct) => action(), new Context(contextData), DefaultCancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="context">Context data that is passed to the exception policy.</param>
+        [DebuggerStepThrough]
+        public void Execute(Action action, Context context)
+        {
+            Execute((ctx, ct) => action(), context, DefaultCancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
+        [DebuggerStepThrough]
+        public void Execute(Action<Context> action, IDictionary<string, object> contextData)
+        {
+            Execute((ctx, ct) => action(ctx), new Context(contextData), DefaultCancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="context">Context data that is passed to the exception policy.</param>
+        [DebuggerStepThrough]
+        public void Execute(Action<Context> action, Context context)
+        {
+            Execute((ctx, ct) => action(ctx), context, DefaultCancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy
+        /// </summary>
+        /// <param name="action"></param>
+        /// <param name="cancellationToken"></param>
+        [DebuggerStepThrough]
+        public void Execute(Action<CancellationToken> action, CancellationToken cancellationToken)
+        {
+            Execute((ctx, ct) => action(ct), new Context(), cancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        [DebuggerStepThrough]
+        public void Execute(Action<CancellationToken> action, IDictionary<string, object> contextData, CancellationToken cancellationToken)
+        {
+            Execute((ctx, ct) => action(ct), new Context(contextData), cancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="context">Context data that is passed to the exception policy.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        [DebuggerStepThrough]
+        public void Execute(Action<CancellationToken> action, Context context, CancellationToken cancellationToken)
+        {
+            Execute((ctx, ct) => action(ct), context, cancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        [DebuggerStepThrough]
+        public void Execute(Action<Context, CancellationToken> action, IDictionary<string, object> contextData, CancellationToken cancellationToken)
+        {
+            Execute(action, new Context(contextData), cancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="context">Context data that is passed to the exception policy.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        [DebuggerStepThrough]
+        public void Execute(Action<Context, CancellationToken> action, Context context, CancellationToken cancellationToken)
+        {
+            if (context == null) throw new ArgumentNullException(nameof(context));
+
+            SetPolicyContext(context);
+
+            ExecuteInternal(action, context, cancellationToken);
+        }
+
+        #region Overloads method-generic in TResult
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the Result.
+        /// </summary>
+        /// <typeparam name="TResult">The type of the Result.</typeparam>
+        /// <param name="action">The action to perform.</param>
+        /// <returns>The value returned by the action</returns>
+        [DebuggerStepThrough]
+        public TResult Execute<TResult>(Func<TResult> action)
+        {
+            return Execute((ctx, ct) => action(), new Context(), DefaultCancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the result.
+        /// </summary>
+        /// <typeparam name="TResult">The type of the result.</typeparam>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        /// <returns>
+        /// The value returned by the action
+        /// </returns>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        [DebuggerStepThrough]
+        public TResult Execute<TResult>(Func<TResult> action, IDictionary<string, object> contextData)
+        {
+            return Execute((ctx, ct) => action(), new Context(contextData), DefaultCancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the result.
+        /// </summary>
+        /// <typeparam name="TResult">The type of the result.</typeparam>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="context">Context data that is passed to the exception policy.</param>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        /// <returns>
+        /// The value returned by the action
+        /// </returns>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        [DebuggerStepThrough]
+        public TResult Execute<TResult>(Func<TResult> action, Context context)
+        {
+            return Execute((ctx, ct) => action(), context, DefaultCancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the result.
+        /// </summary>
+        /// <typeparam name="TResult">The type of the result.</typeparam>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        /// <returns>
+        /// The value returned by the action
+        /// </returns>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        [DebuggerStepThrough]
+        public TResult Execute<TResult>(Func<Context, TResult> action, IDictionary<string, object> contextData)
+        {
+            return Execute((ctx, ct) => action(ctx), new Context(contextData), DefaultCancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the result.
+        /// </summary>
+        /// <typeparam name="TResult">The type of the result.</typeparam>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="context">Context data that is passed to the exception policy.</param>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        /// <returns>
+        /// The value returned by the action
+        /// </returns>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        [DebuggerStepThrough]
+        public TResult Execute<TResult>(Func<Context, TResult> action, Context context)
+        {
+            return Execute((ctx, ct) => action(ctx), context, DefaultCancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the result.
+        /// </summary>
+        /// <typeparam name="TResult">The type of the result.</typeparam>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The value returned by the action</returns>
+        [DebuggerStepThrough]
+        public TResult Execute<TResult>(Func<CancellationToken, TResult> action, CancellationToken cancellationToken)
+        {
+            return Execute((ctx, ct) => action(ct), new Context(), cancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the result.
+        /// </summary>
+        /// <typeparam name="TResult">The type of the result.</typeparam>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The value returned by the action</returns>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        [DebuggerStepThrough]
+        public TResult Execute<TResult>(Func<CancellationToken, TResult> action, IDictionary<string, object> contextData, CancellationToken cancellationToken)
+        {
+            return Execute((ctx, ct) => action(ct), new Context(contextData), cancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the result.
+        /// </summary>
+        /// <typeparam name="TResult">The type of the result.</typeparam>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="context">Context data that is passed to the exception policy.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The value returned by the action</returns>
+        [DebuggerStepThrough]
+        public TResult Execute<TResult>(Func<CancellationToken, TResult> action, Context context, CancellationToken cancellationToken)
+        {
+            return Execute((ctx, ct) => action(ct), context, cancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the result.
+        /// </summary>
+        /// <typeparam name="TResult">The type of the result.</typeparam>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The value returned by the action</returns>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        [DebuggerStepThrough]
+        public TResult Execute<TResult>(Func<Context, CancellationToken, TResult> action, IDictionary<string, object> contextData, CancellationToken cancellationToken)
+        {
+            return Execute(action, new Context(contextData), cancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the result.
+        /// </summary>
+        /// <typeparam name="TResult">The type of the result.</typeparam>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="context">Context data that is passed to the exception policy.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The value returned by the action</returns>
+        [DebuggerStepThrough]
+        public TResult Execute<TResult>(Func<Context, CancellationToken, TResult> action, Context context, CancellationToken cancellationToken)
+        {
+            if (context == null) throw new ArgumentNullException(nameof(context));
+
+            SetPolicyContext(context);
+
+            return ExecuteInternal(action, context, cancellationToken);
+        }
+
+        #endregion
+
+        #endregion
+
+        #region ExecuteAndCapture overloads
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the captured result
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <returns>The captured result</returns>
+        [DebuggerStepThrough]
+        public PolicyResult ExecuteAndCapture(Action action)
+        {
+            return ExecuteAndCapture((ctx, ct) => action(), new Context(), DefaultCancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the captured result.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        /// <returns>The captured result</returns>
+        [DebuggerStepThrough]
+        public PolicyResult ExecuteAndCapture(Action action, IDictionary<string, object> contextData)
+        {
+            return ExecuteAndCapture((ctx, ct) => action(), new Context(contextData), DefaultCancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the captured result.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="context">Context data that is passed to the exception policy.</param>
+        /// <returns>The captured result</returns>
+        [DebuggerStepThrough]
+        public PolicyResult ExecuteAndCapture(Action action, Context context)
+        {
+            return ExecuteAndCapture((ctx, ct) => action(), context, DefaultCancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the captured result.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        /// <returns>The captured result</returns>
+        [DebuggerStepThrough]
+        public PolicyResult ExecuteAndCapture(Action<Context> action, IDictionary<string, object> contextData)
+        {
+            return ExecuteAndCapture((ctx, ct) => action(ctx), new Context(contextData), DefaultCancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the captured result.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="context">Context data that is passed to the exception policy.</param>
+        /// <returns>The captured result</returns>
+        [DebuggerStepThrough]
+        public PolicyResult ExecuteAndCapture(Action<Context> action, Context context)
+        {
+            return ExecuteAndCapture((ctx, ct) => action(ctx), context, DefaultCancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the captured result
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The captured result</returns>
+        [DebuggerStepThrough]
+        public PolicyResult ExecuteAndCapture(Action<CancellationToken> action, CancellationToken cancellationToken)
+        {
+            return ExecuteAndCapture((ctx, ct) => action(ct), new Context(), cancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the captured result.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The captured result</returns>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        [DebuggerStepThrough]
+        public PolicyResult ExecuteAndCapture(Action<CancellationToken> action, IDictionary<string, object> contextData, CancellationToken cancellationToken)
+        {
+            return ExecuteAndCapture((ctx, ct) => action(ct), new Context(contextData), cancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the captured result
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="context">Context data that is passed to the exception policy.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The captured result</returns>
+        [DebuggerStepThrough]
+        public PolicyResult ExecuteAndCapture(Action<CancellationToken> action, Context context, CancellationToken cancellationToken)
+        {
+            return ExecuteAndCapture((ctx, ct) => action(ct), context, cancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the captured result.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The captured result</returns>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        [DebuggerStepThrough]
+        public PolicyResult ExecuteAndCapture(Action<Context, CancellationToken> action, IDictionary<string, object> contextData, CancellationToken cancellationToken)
+        {
+            return ExecuteAndCapture(action, new Context(contextData), cancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the captured result
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="context">Context data that is passed to the exception policy.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The captured result</returns>
+        [DebuggerStepThrough]
+        public PolicyResult ExecuteAndCapture(Action<Context, CancellationToken> action, Context context, CancellationToken cancellationToken)
+        {
+            if (_exceptionPolicy == null) throw new InvalidOperationException(
+                "Please use the synchronous-defined policies when calling the synchronous Execute (and similar) methods.");
+            if (context == null) throw new ArgumentNullException(nameof(context));
+
+            try
+            {
+                Execute(action, context, cancellationToken);
+                return PolicyResult.Successful(context);
+            }
+            catch (Exception exception)
+            {
+                return PolicyResult.Failure(exception, GetExceptionType(ExceptionPredicates, exception), context);
+            }
+        }
+
+        #region Overloads method-generic in TResult
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the captured result
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <returns>The captured result</returns>
+        [DebuggerStepThrough]
+        public PolicyResult<TResult> ExecuteAndCapture<TResult>(Func<TResult> action)
+        {
+            return ExecuteAndCapture((ctx, ct) => action(), new Context(), DefaultCancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the captured result.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        /// <returns>The captured result</returns>
+        [DebuggerStepThrough]
+        public PolicyResult<TResult> ExecuteAndCapture<TResult>(Func<TResult> action, IDictionary<string, object> contextData)
+        {
+            return ExecuteAndCapture((ctx, ct) => action(), new Context(contextData), DefaultCancellationToken);
+        }
+
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the captured result.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="context">Context data that is passed to the exception policy.</param>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        /// <returns>The captured result</returns>
+        [DebuggerStepThrough]
+        public PolicyResult<TResult> ExecuteAndCapture<TResult>(Func<TResult> action, Context context)
+        {
+            return ExecuteAndCapture((ctx, ct) => action(), context, DefaultCancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the captured result.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        /// <returns>The captured result</returns>
+        [DebuggerStepThrough]
+        public PolicyResult<TResult> ExecuteAndCapture<TResult>(Func<Context, TResult> action, IDictionary<string, object> contextData)
+        {
+            return ExecuteAndCapture((ctx, ct) => action(ctx), new Context(contextData), DefaultCancellationToken);
+        }
+
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the captured result.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="context">Context data that is passed to the exception policy.</param>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        /// <returns>The captured result</returns>
+        [DebuggerStepThrough]
+        public PolicyResult<TResult> ExecuteAndCapture<TResult>(Func<Context, TResult> action, Context context)
+        {
+            return ExecuteAndCapture((ctx, ct) => action(ctx), context, DefaultCancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the captured result
+        /// </summary>
+        /// <typeparam name="TResult">The type of the t result.</typeparam>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The captured result</returns>
+        public PolicyResult<TResult> ExecuteAndCapture<TResult>(Func<CancellationToken, TResult> action, CancellationToken cancellationToken)
+        {
+            return ExecuteAndCapture((ctx, ct) => action(ct), new Context(), cancellationToken);
+        }
+
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the captured result.
+        /// </summary>
+        /// <typeparam name="TResult">The type of the result.</typeparam>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The captured result</returns>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        [DebuggerStepThrough]
+        public PolicyResult<TResult> ExecuteAndCapture<TResult>(Func<CancellationToken, TResult> action, IDictionary<string, object> contextData, CancellationToken cancellationToken)
+        {
+            return ExecuteAndCapture((ctx, ct) => action(ct), new Context(contextData), cancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the captured result.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="context">Context data that is passed to the exception policy.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The captured result</returns>
+        [DebuggerStepThrough]
+        public PolicyResult<TResult> ExecuteAndCapture<TResult>(Func<CancellationToken, TResult> action, Context context, CancellationToken cancellationToken)
+        {
+            return ExecuteAndCapture((ctx, ct) => action(ct), context, cancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the captured result.
+        /// </summary>
+        /// <typeparam name="TResult">The type of the result.</typeparam>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="contextData">Arbitrary data that is passed to the exception policy.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The captured result</returns>
+        /// <exception cref="System.ArgumentNullException">contextData</exception>
+        [DebuggerStepThrough]
+        public PolicyResult<TResult> ExecuteAndCapture<TResult>(Func<Context, CancellationToken, TResult> action, IDictionary<string, object> contextData, CancellationToken cancellationToken)
+        {
+            return ExecuteAndCapture(action, new Context(contextData), cancellationToken);
+        }
+
+        /// <summary>
+        /// Executes the specified action within the policy and returns the captured result.
+        /// </summary>
+        /// <param name="action">The action to perform.</param>
+        /// <param name="context">Context data that is passed to the exception policy.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The captured result</returns>
+        [DebuggerStepThrough]
+        public PolicyResult<TResult> ExecuteAndCapture<TResult>(Func<Context, CancellationToken, TResult> action, Context context, CancellationToken cancellationToken)
+        {
+            if (_exceptionPolicy == null) throw new InvalidOperationException(
+                "Please use the synchronous-defined policies when calling the synchronous Execute (and similar) methods.");
+            if (context == null) throw new ArgumentNullException(nameof(context));
+
+            try
+            {
+                return PolicyResult<TResult>.Successful(Execute(action, context, cancellationToken), context);
+            }
+            catch (Exception exception)
+            {
+                return PolicyResult<TResult>.Failure(exception, GetExceptionType(ExceptionPredicates, exception), context);
+            }
+        }
+        #endregion
+
+        internal static ExceptionType GetExceptionType(IEnumerable<ExceptionPredicate> exceptionPredicates, Exception exception)
+        {
+            bool isExceptionTypeHandledByThisPolicy = exceptionPredicates.Any(predicate => predicate(exception) != null);
+
+            return isExceptionTypeHandledByThisPolicy
+                ? ExceptionType.HandledByThisPolicy
+                : ExceptionType.Unhandled;
+        }
+
+        #endregion
+
+    }
+}
