+++ src/Polly.Shared/Polly.Shared.projitems
@@ -85,10 +85,10 @@
     <Compile Include="$(MSBuildThisFileDirectory)Fallback\FallbackPolicyAsync.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Fallback\IFallbackPolicy.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)IAsyncPolicy.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)IAsyncPolicyTResult.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)IAsyncPolicy.TResult.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ISyncPolicy.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)IsPolicy.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ISyncPolicyTResult.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ISyncPolicy.TResult.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)NoOp\INoOpPolicy.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)NoOp\NoOpEngineAsync.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)NoOp\NoOpEngine.cs" />
@@ -98,11 +98,16 @@
     <Compile Include="$(MSBuildThisFileDirectory)NoOp\NoOpTResultSyntaxAsync.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)NoOp\NoOpTResultSyntax.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)NoOp\NoOpSyntax.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Policy.Async.ExecuteOverloads.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Policy.Async.TResult.ExecuteOverloads.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Policy.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Policy.ExecuteOverloads.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Policy.HandleSyntax.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Policy.ContextAndKeys.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)PolicyAsync.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)PolicyAsync.TResult.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Policy.Async.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Policy.Async.TResult.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Policy.TResult.ExecuteOverloads.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)PolicyBase.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PolicyBuilder.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PolicyBuilder.OrSyntax.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PolicyResult.cs" />
